Add Deno KeyValue store
This adds a store type to use Deno's KeyValue database. A new option is
added called "withUrl" that interops with the KV store to track either the
user across the app or on a per-route basis. This is helpful if I want to allow
a lot of access to one route, but limit something like auth signups.